reorganize file structure to reduce import complexity and make have easier navigation
