--- src/engine/game_context/nodes/mod.rs
@@ -1,6 +1,0 @@
-pub mod camera;
-pub mod mesh;
-pub mod model;
-pub mod ui;
-pub mod directional_light;
-pub mod empty;
