--- src/engine/renderer/buffers/mod.rs
@@ -1,5 +1,0 @@
-pub mod frame_buffer;
-pub mod index_buffer;
-pub mod vertex_array;
-pub mod vertex_buffer;
-pub mod vertex_buffer_layout;
